+++ inbm/cloudadapter-agent/tests/unit/cloud/client/test_inbs_cloud_client.py
@@ -56,16 +56,18 @@ class TestInbsCloudClient:
         mock_channel.SendNodeUpdateRequest.return_value = "MockResponse"
         inbs_client._grpc_channel = mock_channel
         
-        key = 'test-key'
-        value = '{"job_id": "12345", "status": 200, "message": "Update successful"}'
+        key = 'update'
+        value = '{"status":200, "message":"COMMAND SUCCESSFUL", "jobId":"swupd-4b151b70-c121-4245-873b-5324ac7a3f7a"}'
         
-        # Call the publish_node_update method
-        inbs_client.publish_node_update(key, value)
+        # Call the publish_update method
+        with patch('cloudadapter.cloud.client.inbs_cloud_client.is_valid_json_structure', return_value=True):
+            inbs_client.publish_update(key, value)
 
         # Assert that the gRPC channel's SendNodeUpdate method was called
         mock_channel.SendNodeUpdate.assert_called_once()
-        
-    def test_publish_node_update_failure_no_grpc_channel(self, inbs_client: InbsCloudClient):
+       
+
+    def test_publish_update_failure_no_grpc_channel(self, inbs_client: InbsCloudClient):
         # Ensure that _grpc_channel is None to simulate the channel not being set up
         inbs_client._grpc_channel = None