+++ inbm/dispatcher-agent/tests/unit/test_update_logger.py
@@ -197,7 +197,7 @@ class TestUpdateLogger(TestCase):
         assert mock_run.call_count == 2
         self.assertEqual(granular_log, expected_content)
 
-    @patch('dispatcher.update_logger.detect_os', return_value='TiberOS')
+    @patch('dispatcher.update_logger.detect_os', return_value='tiber')
     def test_save_granular_log_file_sota_with_tiberos_and_log(self, mock_os) -> None:
         self.update_logger.ota_type = "sota"
         self.update_logger._time = datetime.datetime(2024, 7, 3, 1, 50, 55, 935223)